+++ src/csv/colbased_worker.hpp
@@ -256,7 +256,7 @@ public:
           handlers_[column_index_]->process_float(bsd_strtod(token_.begin(), token_.end()));
         }
         else {
-          handlers_[column_index_]->process_integer(fast_atoi<long>(token_.begin(), token_.end()));
+          handlers_[column_index_]->process_integer(fast_atoi<long long>(token_.begin(), token_.end()));
         }
       }
     } else {
@@ -359,7 +359,7 @@ public:
           }
         }
       if (integer_possible) {
-        handlers_[column_index_]->process_integer(fast_atoi<long>(token_.begin(), token_.end()));
+        handlers_[column_index_]->process_integer(fast_atoi<long long>(token_.begin(), token_.end()));
       }
       else if (float_possible || exp_possible) {
         handlers_[column_index_]->process_float(bsd_strtod(token_.begin(), token_.end()));
@@ -391,7 +391,7 @@ private:
   std::vector<std::shared_ptr<ColumnHandler> > handlers_;
   std::vector<char>                            token_;
   std::vector<char>                            token_aux_;
-  std::vector<std::pair<size_t, long> >        long_cache_;
+  std::vector<std::pair<size_t, long long> >   long_cache_;
   std::vector<std::pair<size_t, double> >      double_cache_;
   std::vector<char>                            str_cache_data_;
   std::vector<size_t>                          str_cache_offsets_;
