+++ src/diagnostic/parse_and_sum.hpp
@@ -157,7 +157,7 @@ public:
           return bsd_strtod(begin, end);
         }
         else {
-          return (double)fast_atoi<long>(begin, end);
+          return (double)fast_atoi<long long>(begin, end);
         }
       }
     }
@@ -205,7 +205,7 @@ private:
         chunker_.process(filename, 0, params.num_threads, params.allow_quoted_newlines);
       }
       for (size_t worker_id = 0; worker_id < chunker_.num_chunks(); worker_id++) {
-        long start_of_chunk = 0, end_of_chunk = 0;
+        long long start_of_chunk = 0, end_of_chunk = 0;
         std::tie(start_of_chunk, end_of_chunk) = chunker_.get_chunk(worker_id);
         if (start_of_chunk < 0 || end_of_chunk < 0) {
           continue;
